requirements.txt
Django==2.1.7
amqp==2.1.4
django-allauth==0.37.1
django-debug-toolbar==1.8
django_extensions==2.1.5
djangorestframework==3.8.2
gunicorn==19.6.0
mysqlclient==1.3.13